Module mercury::header

Structs

ConnectionBuffer

ConnectionID
ID to keep track of clients

QuicSocket
Custom socket which can hold client address info and buffer for data specific to that connection

TlsBuffer
Buffer with custom read/write trait implementations for TLS messages

Enums

Header
Format of messages which will be sent between client and server

HeaderType
Two Header types which can be sent

PacketNumber
How many octets are being used for the packet number section of a ShortHeader packet

PacketTypeLong
Type of LongHeader packet

PacketTypeShort
How many octets are being used for the packet number section of a ShortHeader packet

Functions

connection_id_omitted
This function is primarily for readability - returns the bool given for connection_omit by get_short_info

connection_id_present
Used for getting a straightforward answer to whether a Connection ID is present

decode
Reconstruct a Header from Bytes

get_long_info
Parse info from first octet of LongHeader - currently this only consists of PacketTypeLong

get_short_info
Get information for ShortHeader packet