Enum mercury::header::Header

pub enum Header {
    LongHeader {
        packet_type: PacketTypeLong,
        connection_id: u64,
        packet_number: u32,
        version: u32,
        payload: Vec<u8>,
    },
    ShortHeader {
        key_phase: u8,
        connection_id: Option<ConnectionID>,
        packet_number: PacketNumber,
        payload: Vec<u8>,
    },
}

Format of messages which will be sent between client and server

Variants
LongHeader
Fields of LongHeader

packet_type: PacketTypeLong

connection_id: u64

packet_number: u32

version: u32

payload: Vec<u8>

ShortHeader
Fields of ShortHeader

key_phase: u8

connection_id: Option<ConnectionID>

packet_number: PacketNumber

payload: Vec<u8>

Methods

impl Header

fn encode(self) -> Bytes

Return a representation of Header as Bytes

fn is_new_connection(&self) -> bool

DEPRECATED: NO LONGER USED

fn is_compatible_version(&self) -> bool

DEPRECATED: DO NOT USE

fn get_conn_id(&self) -> Option<u64>

Obtain a Header's packet number, if it is present

fn get_packet_number(&self) -> u32

Obtain a Header's packet number

Note that OneOctet(u8) and TwoOctet(u16) packet numbers will be cast to u32

fn get_version(&self) -> Option<u32>

Obtain a Header's version number, if present

Will always return None for ShortHeader variants

fn get_payload(&self) -> &Vec<u8>

Obtain an immutable reference to a Header's payload - payload data should not be modified within the struct, only read by other functions